(fix) O3-5763: Expose the visit notes form as a reusable workspace
