+++ packages/esm-patient-notes-app/src/index.ts
@@ -32,6 +32,11 @@ export const visitNotesActionButton = getSyncLifecycle(visitNotesActionButtonExt
 
 // t('visitNoteWorkspaceTitle', 'Visit Note')
 export const visitNotesFormWorkspace = getAsyncLifecycle(() => import('./notes/visit-notes-form.workspace'), options);
+
+export const exportedVisitNotesFormWorkspace = getAsyncLifecycle(
+  () => import('./notes/exported-visit-notes-form.workspace'),
+  options,
+);
 export const stickyNotesButton = getAsyncLifecycle(
   () => import('./sticky-notes/sticky-note-header-button.component'),
   options,
